requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-decouple>=3.8
Pillow>=10.0
psycopg2-binary>=2.9.7
django-storages>=1.14.0
boto3>=1.28.0
gunicorn>=21.2.0
whitenoise>=6.5.0
django-htmx>=1.23.0
django-components>=0.83
django-webpack-loader>=2.0.0
django-allauth>=0.57.0
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
supabase>=1.0.3
python-dotenv>=1.0.0
stripe>=7.0.0
requests>=2.31.0
django-csp>=3.7